backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==0.28.0
requests==2.31.0 